requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.10.1
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jsonschema==4.2.0
MarkupSafe==2.0.1
mysqlclient==2.0.3
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.2
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.1.2
requests==2.26.0
requests-oauthlib==1.3.0
rest-swagger==1.0.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
